Cambios en tipos: normalizado todo a Double, y corregidos algunos tipos de funciones (eran tipos simples y era necesario que fuesen collections o streams) --- Sacados los errores feos que agrego eclipse (import Examen, wtf)

+++ DDS-correctorParciales/src/correctorParciales/MetodoCorreccion.java
@@ -1,9 +1,7 @@
 package correctorParciales;
-import Examen;
-
 
 public interface MetodoCorreccion {
 	
-	public float notaParaExamen(Examen unExamen);
+	public double notaParaExamen(Examen unExamen);
 
 }
