+++ DDS-correctorParciales/src/correctorParciales/ParDeConversion.java
@@ -2,30 +2,30 @@ package correctorParciales;
 
 public class ParDeConversion {
 	
-	private float puntaje;
-	private float notaCorrespondiente;
+	private double puntaje;
+	private double notaCorrespondiente;
 	
-	public void setPuntaje(float unPuntaje){
+	public void setPuntaje(double unPuntaje){
 		this.puntaje=unPuntaje;
 	}
 	
-	public float getPuntaje(){
+	public double getPuntaje(){
 		return this.puntaje;
 	}
 	
-	public void setNotaCorrespondiente(float unaNota){
+	public void setNotaCorrespondiente(double unaNota){
 		this.notaCorrespondiente=unaNota;
 	}
 	
-	public float getNotaCorrespondiente(){
+	public double getNotaCorrespondiente(){
 		return this.notaCorrespondiente;
 	}
 	
-	public boolean coincideCon(float unPuntaje){
+	public boolean coincideCon(double unPuntaje){
 		return (this.getPuntaje() == unPuntaje);
 	}
 	
-	public ParDeConversion(float unPuntaje, float unaNota){
+	public ParDeConversion(double unPuntaje, double unaNota){
 		this.puntaje=unPuntaje;
 		this.notaCorrespondiente=unaNota;
 	}
