+++ DDS-correctorParciales/src/correctorParciales/ReglaDeTres.java
@@ -1,10 +1,8 @@
 package correctorParciales;
-import Examen;
-
 
 public class ReglaDeTres implements MetodoCorreccion {
 	
-	public float notaParaExamen(Examen unExamen) {
+	public double notaParaExamen(Examen unExamen) {
 		return (unExamen.puntajeTotal())*10/(unExamen.getRespuestasDadas().size());
 	}
